Squashed commit of the following:
commit 50f48b80c5504eb9c9ba066a3137be2df533a6d3
Author: Dayanfreitas <dayan.freitas@catolicasc.org.edu>
Date:   Sun Nov 19 17:40:22 2023 -0300

    feat: :sparkles: template desktop

commit 198f211d23bea044de7d18f75018bbdd4825a2df
Author: Dayanfreitas <dayan.freitas@catolicasc.org.edu>
Date:   Sun Nov 19 17:17:30 2023 -0300

    refactor: :recycle: Movido footer para dentro do template root

commit f04459425db9b19eb96be0b91a5eef0c024bfd91
Author: Dayanfreitas <dayan.freitas@catolicasc.org.edu>
Date:   Sun Nov 19 17:16:35 2023 -0300

    feat: :sparkles: created root template
